Pull duplicate logic into separate file
